+++ e2e-tests/contentful/src/components/layout.js
@@ -22,7 +22,7 @@ const Layout = ({ children }) => (
           style={{
             margin: `0 auto`,
             maxWidth: 960,
-            padding: `0px 1.0875rem 1.45rem`,
+            padding: `0px 20px`,
             paddingTop: 0,
           }}
         >