+++ packages/gatsby/src/services/initialize.ts
@@ -26,6 +26,7 @@ import { internalActions } from "../redux/actions"
 import { IGatsbyState } from "../redux/types"
 import { IBuildContext } from "./types"
 import availableFlags from "../utils/flags"
+import { detectLmdbStore } from "../datastore"
 
 interface IPluginResolution {
   resolve: string
@@ -204,6 +205,7 @@ export async function initialize({
     // enable loading indicator
     process.env.GATSBY_QUERY_ON_DEMAND_LOADING_INDICATOR = `true`
   }
+  detectLmdbStore()
 
   // theme gatsby configs can be functions or objects
   if (config) {