+++ packages/gatsby/src/utils/nodes.ts
@@ -1,5 +1,39 @@
+import { createPageDependency } from "../redux/actions/add-page-dependency"
+import { getNode } from "../datastore"
 import { IGatsbyNode } from "../redux/types"
 import { store } from "../redux"
+
+/**
+ * Determine if node has changed.
+ */
+export const hasNodeChanged = (id: string, digest: string): boolean => {
+  const node = getNode(id)
+  if (!node) {
+    return true
+  } else {
+    return node.internal.contentDigest !== digest
+  }
+}
+
+/**
+ * Get node and save path dependency.
+ */
+export const getNodeAndSavePathDependency = (
+  id: string,
+  path: string
+): IGatsbyNode | undefined => {
+  const node = getNode(id)
+
+  if (!node) {
+    console.error(
+      `getNodeAndSavePathDependency failed for node id: ${id} as it was not found in cache`
+    )
+    return undefined
+  }
+
+  createPageDependency({ path, nodeId: id })
+  return node
+}
 
 /**
  * Get content for a node from the plugin that created it.