+++ packages/gatsby/src/utils/source-nodes.ts
@@ -2,7 +2,7 @@ import report from "gatsby-cli/lib/reporter"
 import { Span } from "opentracing"
 import apiRunner from "./api-runner-node"
 import { store } from "../redux"
-import { getNode, getNodes } from "../redux/nodes"
+import { getDataStore, getNode, getNodes } from "../datastore"
 import { actions } from "../redux/actions"
 import { IGatsbyState } from "../redux/types"
 const { deleteNode } = actions
@@ -104,6 +104,7 @@ export default async ({
     webhookBody: webhookBody || {},
     pluginName,
   })
+  await getDataStore().ready()
 
   const state = store.getState()
   const nodes = getNodes()