+++ packages/gatsby-plugin-netlify/CHANGELOG.md
@@ -3,6 +3,16 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [3.7.0-next.2](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-netlify@3.7.0-next.1...gatsby-plugin-netlify@3.7.0-next.2) (2021-06-02)
+
+**Note:** Version bump only for package gatsby-plugin-netlify
+
+# [3.7.0-next.1](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-netlify@3.7.0-next.0...gatsby-plugin-netlify@3.7.0-next.1) (2021-05-31)
+
+### Bug Fixes
+
+- **gatsby:** add componentChunkName to components list so don't need to loop over pages ([#31547](https://github.com/gatsbyjs/gatsby/issues/31547)) ([783b937](https://github.com/gatsbyjs/gatsby/commit/783b937c8f70478796bce37808bf8bf967bb4252))
+
 # [3.7.0-next.0](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-netlify@3.6.0-next.0...gatsby-plugin-netlify@3.7.0-next.0) (2021-05-21)
 
 **Note:** Version bump only for package gatsby-plugin-netlify