+++ packages/gatsby-source-lever/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [3.7.0-next.2](https://github.com/gatsbyjs/gatsby/compare/gatsby-source-lever@3.7.0-next.1...gatsby-source-lever@3.7.0-next.2) (2021-06-02)
+
+**Note:** Version bump only for package gatsby-source-lever
+
+# [3.7.0-next.1](https://github.com/gatsbyjs/gatsby/compare/gatsby-source-lever@3.7.0-next.0...gatsby-source-lever@3.7.0-next.1) (2021-05-31)
+
+**Note:** Version bump only for package gatsby-source-lever
+
 # [3.7.0-next.0](https://github.com/gatsbyjs/gatsby/compare/gatsby-source-lever@3.6.0-next.0...gatsby-source-lever@3.7.0-next.0) (2021-05-21)
 
 **Note:** Version bump only for package gatsby-source-lever